backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
requests
python-decouple
python-dotenv
django-cors-headers
